+++ Modules/Microsoft365DSC/Examples/Resources/AADEntitlementManagementRoleAssignment/3-Remove.ps1
@@ -29,7 +29,7 @@ Configuration Example
             TenantId              = $TenantId
             CertificateThumbprint = $CertificateThumbprint
             Ensure          = "Absent";
-            Principal       = "John.Smith@$TenantId";
+            Principal       = "AdeleV@$TenantId";
             RoleDefinition  = "Catalog creator";
         }
     }