+++ Modules/Microsoft365DSC/Examples/Resources/EXOAvailabilityConfig/3-Remove.ps1
@@ -5,22 +5,31 @@ It is not meant to use as a production baseline.
 
 Configuration Example
 {
-    param
-    (
-        [Parameter(Mandatory = $true)]
-        [PSCredential]
-        $Credscredential
+    param(
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint
     )
+
     Import-DscResource -ModuleName Microsoft365DSC
 
-    $Domain = $Credscredential.Username.Split('@')[1]
     node localhost
     {
         EXOAvailabilityConfig 'ConfigureAvailabilityConfig'
         {
-            OrgWideAccount       = "alexW@$Domain" # Updated Property
+            OrgWideAccount       = "alexW@$TenantId" # Updated Property
             Ensure               = "Absent"
-            Credential           = $Credscredential
+            ApplicationId         = $ApplicationId
+            TenantId              = $TenantId
+            CertificateThumbprint = $CertificateThumbprint
         }
     }
 }