+++ Modules/Microsoft365DSC/Examples/Resources/EXOCASMailboxSettings/2-Update.ps1
@@ -5,16 +5,21 @@ It is not meant to use as a production baseline.
 
 Configuration Example
 {
-    param
-    (
-        [Parameter(Mandatory = $true)]
-        [PSCredential]
-        $Credscredential
+    param(
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint
     )
 
     Import-DscResource -ModuleName Microsoft365DSC
-
-    $Domain = $Credscredential.Username.Split('@')[1]
     node localhost
     {
         EXOCASMailboxSettings 'AdeleVCasMailboxSettings'
@@ -26,7 +31,7 @@ Configuration Example
             ActiveSyncMailboxPolicy                 = 'Default'
             ActiveSyncSuppressReadReceipt           = $False
             EwsEnabled                              = $True
-            Identity                                = "admin@$Domain"
+            Identity                                = "admin@$TenantId"
             ImapEnabled                             = $True # Updated Property
             ImapForceICalForCalendarRetrievalOption = $False
             ImapMessagesRetrievalMimeFormat         = 'BestBodyFormat'
@@ -47,7 +52,9 @@ Configuration Example
             ShowGalAsDefaultView                    = $True
             UniversalOutlookEnabled                 = $True
             Ensure                                  = 'Present'
-            Credential                              = $Credscredential
+            ApplicationId         = $ApplicationId
+            TenantId              = $TenantId
+            CertificateThumbprint = $CertificateThumbprint
         }
     }
 }