+++ Modules/Microsoft365DSC/Examples/Resources/EXOManagementRole/1-Create.ps1
@@ -6,22 +6,31 @@ It is not meant to use as a production baseline.
 Configuration Example
 {
     param(
-        [Parameter(Mandatory = $true)]
-        [PSCredential]
-        $Credscredential
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
-    $Domain = $Credscredential.Username.Split('@')[1]
     node localhost
     {
         EXOManagementRole 'ConfigureManagementRole'
         {
             Name                 = "MyDisplayName"
             Description          = ""
-            Parent               = "$Domain\MyProfileInformation"
+            Parent               = "$TenantId\MyProfileInformation"
             Ensure               = "Present"
-            Credential           = $Credscredential
+            ApplicationId         = $ApplicationId
+            TenantId              = $TenantId
+            CertificateThumbprint = $CertificateThumbprint
         }
     }
 }