+++ Modules/Microsoft365DSC/Examples/Resources/EXOQuarantinePolicy/3-Remove.ps1
@@ -6,9 +6,17 @@ It is not meant to use as a production baseline.
 Configuration Example
 {
     param(
-        [Parameter(Mandatory = $true)]
-        [PSCredential]
-        $Credscredential
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint
     )
     $OrganizationName = $Credscredential.UserName.Split('@')[1]
     Import-DscResource -ModuleName Microsoft365DSC
@@ -18,9 +26,11 @@ Configuration Example
     {
         EXOQuarantinePolicy 'ConfigureQuarantinePolicy'
         {
-            Identity                          = "$Domain\IntegrationPolicy";
+            Identity                          = "$TenantId\IntegrationPolicy";
             Ensure                            = "Absent"
-            Credential                        = $Credscredential
+            ApplicationId         = $ApplicationId
+            TenantId              = $TenantId
+            CertificateThumbprint = $CertificateThumbprint
         }
     }
 }