+++ CHANGELOG.md
@@ -9,6 +9,10 @@
   * Fixes an issue with nested and duplicate settings in the settings catalog
 * M365DSCResourceGenerator
   * Update Intune resource generation.
+* M365DSCReport
+  * Changes behaviour to not throw on empty configuration during report generation.  
+    FIXES [#4559](https://github.com/microsoft/Microsoft365DSC/issues/4559)  
+    FIXES [#4505](https://github.com/microsoft/Microsoft365DSC/issues/4505)
 
 # 1.24.710.3